pyproject.toml
[project]
name = "pilot-em-agent"
version = "0.2.0"
description = "Pilot: NL -> openEMS FDTD agent (free stack, mock fallback)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.27",
    "httpx>=0.27",
    "pypdf>=4.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[tool.pytest.ini_options]
testpaths = ["tests"]
